requirements.txt
# Telegram
python-telegram-bot==20.7
telethon==1.25.6
python-dotenv==1.0.1
httpx==0.28.1

# Database
sqlalchemy==2.0.23

# LLM
llama-cpp-python==0.2.90

# Optional dependencies
Pillow==10.1.0
openai==1.3.7
google-generativeai==0.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
